+++ chatbot/patient/cancelAppointment.ts
@@ -1,8 +1,8 @@
-import { PatientState, TrxOrDb } from '../types.ts'
+import { PatientState, TrxOrDb } from '../../types.ts'
 import { assert } from 'std/testing/asserts.ts'
-import * as google from '../external-clients/google.ts'
-import { getWithTokensById } from '../db/models/health_workers.ts'
-import { deleteAppointment } from '../db/models/appointments.ts'
+import * as google from '../../external-clients/google.ts'
+import { getWithTokensById } from '../../db/models/health_workers.ts'
+import { deleteAppointment } from '../../db/models/appointments.ts'
 
 // This should remove the scheduled appointment from the database and from google calendar
 export async function cancelAppointment(
